accruals service to calculate the emi,penalty,interest,latefee,overdue using scheduler and fetching the totalaccuralsbyloannumber from bankservice using feignclient

+++ bank-service/src/main/java/com/rashmita/bankservice/model/LoanUpdateRequest.java
@@ -13,8 +13,8 @@ public class LoanUpdateRequest {
     private Long id;
     private int minimumAmount;
     private int maximumAmount;
-    private Double interestRate;
-    private Double lateFeeCharge;
+    private int interestRate;
+    private double lateFeeCharge;
     private Double loanAdministrationFeeRate;
     private Double loanAdministrationFeeAmount;
     private int defaultingPeriod;
